Implemented User Authentication
Added User Authentication tokens that are created when a user logs in or registers and are stored in the session storage. Also added password hashing and reworked the frontend RegisterScreen and LoginScreen to hold the register and login logic within those files as opposed to their respective button components.

Also added the Database DROP and CREATE queries which should automatically configure an empty database so long as it exists and a clash_admin user has permission to access it.