Implemented Logout and storing data in sessionStorage to improve load times.  Added Devs as users to be automatically added to the database.
Logout is accessible in profile screen and clears both local and session storage.
Data loaded from the canvas API is stored in sessionStorage and reloaded upon a refresh.
Usernames for devs are fLast where f is your first initial and Last is your last name, with the initial letter capitalized. Password is 'pass'.